requirements.txt
asgiref==3.2.3
Django==3.0.3
django-colorfield==0.3.0
mimesis==4.0.0
pkg-resources==0.0.0
python-decouple==3.3
pytils==0.3
pytz==2019.3
sqlparse==0.3.0
